Session handling tweaks and read-only guest access

+++ includes/functions.php
@@ -17,7 +17,9 @@ function getListRow_Service($row) {
 				echo '<td>';
 					echo '<img src="' . $row['RETURN_ARG7'] . '" class="list_art"/>';
 					echo '<div class="list_info">';
-						echo '<a href="includes/delete_song.php?id=' . $row['TRACKID'] . '&user=' . $_SESSION['username'] . '"/><span class="oi" data-glyph="delete" id="list_delete"></span></a>';
+						if ($_SESSION['login']) {
+							echo '<a href="includes/delete_song.php?id=' . $row['TRACKID'] . '&user=' . $_SESSION['username'] . '"/><span class="oi" data-glyph="delete" id="list_delete"></span></a>';
+						}
 						echo '<span class="list_title">' . $row['RETURN_ARG2'] . '</span><br/>';
 						echo '<span class="list_artist"><a href="' . $row['RETURN_ARG4'] . '">' . $row['RETURN_ARG3'] . '</a></span>';
 					echo '</div>';
@@ -32,7 +34,9 @@ function getListRow_Service($row) {
 				echo '<td>';
 					echo '<img src="' . $row['RETURN_ARG7'] . '" class="list_art"/>';
 					echo '<div class="list_info">';
-						echo '<a href="includes/delete_song.php?id=' . $row['TRACKID'] . '&user=' . $_SESSION['username'] . '"/><span class="oi" data-glyph="delete" id="list_delete"></span></a>';
+						if ($_SESSION['login']) {
+							echo '<a href="includes/delete_song.php?id=' . $row['TRACKID'] . '&user=' . $_SESSION['username'] . '"/><span class="oi" data-glyph="delete" id="list_delete"></span></a>';
+						}
 						echo '<span class="list_title">' . $row['RETURN_ARG2'] . '</span><br/>';
 						echo '<span class="list_artist"><a href="' . $row['RETURN_ARG4'] . '">' . $row['RETURN_ARG3'] . '</a></span>';
 					echo '</div>';
